+++ Tests/ToolBox/RandomDataGenerator/OptionSymbolGeneratorTests.cs
@@ -1,0 +1,139 @@
+/*
+ * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
+ * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+using Moq;
+using NUnit.Framework;
+using QuantConnect.ToolBox.RandomDataGenerator;
+using System;
+using System.Linq;
+
+namespace QuantConnect.Tests.ToolBox.RandomDataGenerator
+{
+    [TestFixture]
+    public class OptionSymbolGeneratorTests
+    {
+        private const int Seed = 123456789;
+        private static readonly IRandomValueGenerator _randomValueGenerator = new RandomValueGenerator(Seed);
+
+        private BaseSymbolGenerator _symbolGenerator;
+        private DateTime _minExpiry = new(2000, 01, 01);
+        private DateTime _maxExpiry = new(2001, 01, 01);
+        private decimal _underlyingPrice = 100m;
+        private decimal _maximumStrikePriceDeviation = 50m;
+
+        [SetUp]
+        public void Setup()
+        {
+            // initialize using a seed for deterministic tests
+            _symbolGenerator = new OptionSymbolGenerator(
+                new RandomDataGeneratorSettings()
+                {
+                    Market = Market.USA,
+                    Start = _minExpiry,
+                    End = _maxExpiry
+                },
+                _randomValueGenerator,
+                _underlyingPrice,
+                _maximumStrikePriceDeviation);
+        }
+
+        [Test]
+        [TestCase(SecurityType.Option)]
+        public void ReturnsFutureSymbolGeneratorInstance(SecurityType securityType)
+        {
+            Assert.IsInstanceOf<OptionSymbolGenerator>(BaseSymbolGenerator.Create(
+                new RandomDataGeneratorSettings { SecurityType = securityType },
+                Mock.Of<IRandomValueGenerator>()
+            ));
+        }
+
+        [Test]
+        public void GetAvailableSymbolCount()
+        {
+            Assert.AreEqual(int.MaxValue,
+                new OptionSymbolGenerator(Mock.Of<RandomDataGeneratorSettings>(), Mock.Of<RandomValueGenerator>(), 100m,
+                    75m).GetAvailableSymbolCount());
+        }
+
+        [Test]
+        public void NextOptionSymbol_CreatesOptionSymbol_WithCorrectSecurityTypeAndEquityUnderlying()
+        {
+            var securities = _symbolGenerator.GenerateAsset().ToList();
+            Assert.AreEqual(2, securities.Count);
+
+            var underlying = securities[0];
+            var option = securities[1];
+
+            Assert.AreEqual(SecurityType.Option, option.SecurityType);
+
+            var underlyingOrigin = option.Underlying;
+            Assert.AreEqual(underlying.Value, underlyingOrigin.Value);
+            Assert.AreEqual(Market.USA, underlying.ID.Market);
+            Assert.AreEqual(SecurityType.Equity, underlying.SecurityType);
+        }
+
+        [Test]
+        public void NextOptionSymbol_CreatesOptionSymbol_WithinSpecifiedExpiration_OnFriday()
+        {
+            var securities = _symbolGenerator.GenerateAsset().ToList();
+            Assert.AreEqual(2, securities.Count);
+
+            var option = securities[1];
+
+            var expiration = option.ID.Date;
+            Assert.LessOrEqual(_minExpiry, expiration);
+            Assert.GreaterOrEqual(_maxExpiry, expiration);
+        }
+
+        [Test]
+        public void NextOptionSymbol_CreatesOptionSymbol_WithRequestedMarket()
+        {
+            for (int i = 0; i < 50; i++)
+            {
+                var price = _randomValueGenerator.NextPrice(SecurityType.Equity, Market.USA, 100m, 100m);
+                var symbolGenerator = new OptionSymbolGenerator(
+                    new RandomDataGeneratorSettings()
+                    {
+                        Market = Market.USA,
+                        Start = _minExpiry,
+                        End = _maxExpiry
+                    },
+                    _randomValueGenerator,
+                    price,
+                    50m);
+                var securities = symbolGenerator.GenerateAsset().ToList();
+                Assert.AreEqual(2, securities.Count);
+
+                var option = securities[1];
+
+                Assert.AreEqual(Market.USA, option.ID.Market);
+            }
+        }
+
+        [Test]
+        public void NextOptionSymbol_CreatesOptionSymbol_WithinSpecifiedStrikePriceDeviation()
+        {
+            var securities = _symbolGenerator.GenerateAsset().ToList();
+            Assert.AreEqual(2, securities.Count);
+
+            var option = securities[1];
+
+            var strikePrice = option.ID.StrikePrice;
+            var maximumDeviation = _underlyingPrice * (_maximumStrikePriceDeviation / 100m);
+            Assert.LessOrEqual(_underlyingPrice - maximumDeviation, strikePrice);
+            Assert.GreaterOrEqual(_underlyingPrice + maximumDeviation, strikePrice);
+        }
+    }
+}
