+++ Tests/ToolBox/RandomDataGenerator/RandomPriceGeneratorTests.cs
@@ -1,0 +1,51 @@
+/*
+ * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
+ * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+using Moq;
+using NUnit.Framework;
+using QuantConnect.ToolBox.RandomDataGenerator;
+using System;
+
+namespace QuantConnect.Tests.ToolBox.RandomDataGenerator
+{
+    [TestFixture]
+    public class RandomPriceGeneratorTests
+    {
+        [Test]
+        public void ReturnRandomReference()
+        {
+            var randomMock = new Mock<IRandomValueGenerator>();
+            decimal expect = 42m;
+            randomMock
+                .Setup(s => s.NextPrice(It.IsAny<SecurityType>(), It.IsAny<string>(), It.IsAny<decimal>(), It.IsAny<decimal>()))
+                .Returns(expect);
+            var randomPriceGenerator = new RandomPriceGenerator(Symbols.AAPL, randomMock.Object);
+
+            var actual = randomPriceGenerator.NextReferencePrice(10, 20);
+            randomMock.Verify(s => s.NextPrice(SecurityType.Equity, Market.USA, 10, 20), Times.Once);
+            Assert.AreEqual(expect, actual);
+        }
+
+        [Test]
+        public void ReturnSameAsReference()
+        {
+            var randomMock = new Mock<IRandomValueGenerator>();
+            var randomPriceGenerator = new RandomPriceGenerator(Symbols.AAPL, randomMock.Object);
+
+            var actual = randomPriceGenerator.NextValue(10, DateTime.MinValue);
+            Assert.AreEqual(10, actual);
+        }
+    }
+}
